+++ front/app/components/LandingPages/admin/CTARadioButtons.tsx
@@ -9,7 +9,7 @@ import { TCustomPageCTAType } from 'api/custom_pages/types';
 import {
   CTASignedInType,
   CTASignedOutType,
-} from 'containers/Admin/pagesAndMenu/containers/ContentBuilder/components/CraftComponents/HomepageBanner';
+} from 'containers/Admin/pagesAndMenu/containers/ContentBuilder/components/Widgets/HomepageBanner';
 
 import Error, { TFieldName } from 'components/UI/Error';
 import InputMultilocWithLocaleSwitcher from 'components/UI/InputMultilocWithLocaleSwitcher';
