+++ front/app/containers/Admin/pagesAndMenu/containers/ContentBuilder/components/Widgets/HomepageBanner/LayoutSettingField/LayoutSettingField.test.tsx
