+++ front/app/containers/Admin/pagesAndMenu/containers/ContentBuilder/components/Widgets/OpenToParticipation/ProjectCarrousel/LightProjectCard.tsx
