+++ front/app/containers/Admin/pagesAndMenu/containers/ContentBuilder/components/Widgets/Spotlight/index.tsx
