+++ front/app/containers/Admin/pagesAndMenu/containers/ContentBuilder/components/Widgets/_deprecated/Proposals/index.tsx
