+++ front/app/containers/HomePage/SignedOutHeader/CTA.tsx
@@ -2,7 +2,7 @@ import React, { KeyboardEvent, MouseEvent } from 'react';
 
 import useLocalize from 'hooks/useLocalize';
 
-import { IHomepageBannerSettings } from 'containers/Admin/pagesAndMenu/containers/ContentBuilder/components/CraftComponents/HomepageBanner';
+import { IHomepageBannerSettings } from 'containers/Admin/pagesAndMenu/containers/ContentBuilder/components/Widgets/HomepageBanner';
 import SignUpButton from 'containers/HomePage/SignUpButton';
 
 import BannerButton, {
