+++ front/app/containers/HomePage/SignedOutHeader/FixedRatioLayout.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { IHomepageBannerSettings } from 'containers/Admin/pagesAndMenu/containers/ContentBuilder/components/CraftComponents/HomepageBanner';
+import { IHomepageBannerSettings } from 'containers/Admin/pagesAndMenu/containers/ContentBuilder/components/Widgets/HomepageBanner';
 
 import {
   Container,
