+++ front/app/containers/HomePage/SignedOutHeader/SignedOutHeader.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { IHomepageBannerSettings } from 'containers/Admin/pagesAndMenu/containers/ContentBuilder/components/CraftComponents/HomepageBanner';
+import { IHomepageBannerSettings } from 'containers/Admin/pagesAndMenu/containers/ContentBuilder/components/Widgets/HomepageBanner';
 
 import { render, screen } from 'utils/testUtils/rtl';
